move courseEventsDelete to own files
